我的界面

+++ app/src/main/java/com/anfly/anflyshop/model/bean/TopicDetailBean.java
@@ -66,7 +66,7 @@ public class TopicDetailBean {
         private String item_pic_url;
         private String subtitle;
         private int topic_category_id;
-        private int price_info;
+        private float price_info;
         private String read_count;
         private String scene_pic_url;
         private int topic_template_id;
@@ -130,11 +130,11 @@ public class TopicDetailBean {
             this.topic_category_id = topic_category_id;
         }
 
-        public int getPrice_info() {
+        public float getPrice_info() {
             return price_info;
         }
 
-        public void setPrice_info(int price_info) {
+        public void setPrice_info(float price_info) {
             this.price_info = price_info;
         }
